+++ Velad/Source/VLDWeekTableViewCell.m
@@ -28,7 +28,7 @@
 - (void)setViewModel:(VLDWeekViewModel *)viewModel {
     _viewModel = viewModel;
     self.textLabel.text = viewModel.basicPoint.name;
-    self.detailTextLabel.text = [NSString stringWithFormat:@"%d/7", viewModel.weekCount];
+    self.detailTextLabel.text = [NSString stringWithFormat:@"%ld/7", (long)viewModel.weekCount];
 }
 
 @end
